dbotapi/pyproject.toml
[project]
name = "dbotapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "jwt>=1.3.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "requests>=2.32.3",
]
